refactored js chart
