+++ app/assets/stylesheets/components/_calculator.scss
@@ -3,16 +3,6 @@
   padding: 0;
 }
 
-.color {
-  height: 330px;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-}
-
-
 .button {
   width: 50px;
   height: 50px;
@@ -55,6 +45,19 @@
   // top: 50%;
   // transform: translateX(-50%) translateY(-50%);
 
+}
+
+.color {
+  height: 390px;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+@media (min-width: 768px) {
+  .color {
+    height: 330px;
+  }
 }
