+++ app/assets/stylesheets/components/_utilities.scss
@@ -89,5 +89,8 @@ a:hover {
   height: 20px;
 }
 .padding-15 {
-  padding: 10px 0;
+  padding: 15px 0;
+}
+.padding-30 {
+  padding: 30px;
 }
