+++ app/views/videos/_calculater.html.erb
@@ -1,7 +1,7 @@
 <div class="color">
   <div class="main">
     <form name="form">
-      <input class="textview" name="textview">
+      <input class="textview form-control" name="textview">
     </form>
     <table>
       <tr>
