+++ app/views/videos/show.html.erb
@@ -1,56 +1,55 @@
-<h1 class="text-center headline"><!--<%= @video.title %>--></h1>
-
+<div class="padding-30"></div>
 <div class="container">
   <div class="grid-show">
     <div class="wrapper-row">
       <div class="row">
         <div class="col-xs-12 col-md-12">
-          <div class="iframe-container">
-            <iframe src="https://www.youtube.com/embed/<%= @video.youtube_id %>?rel=0" allowfullscreen></iframe>
-          </div>
-            <div class="bottom-video">
-              <div class="video-details-wrapper">
-
-                <p><i class="far fa-comments"> 12</i></p>
-                <p><i class="far fa-eye"> 2</i></p>
-                <div class="icon-wrapper"><i class="far fa-clock"></i> <%= @video.duration_seconds / 60 %>:<%= "%.2i" %(@video.duration_seconds % 60) %>
-                </div>
-              </div>
-              <div class="verified-wrapper">
-                <% if @video.verified == true  %>
-                  <div class="verified">
-                    <i class="fas fa-check-circle"><p>Vidéo vérifiée</p></i>
-                  </div>
-                <% end %>
-                <% if @video.introduction == true  %>
-                  <div class="verified">
-                    <i class="fas fa-info-circle"><p>Vidéo d'introduction</p></i>
-
-                  </div>
-                <% end %>
-                <% if  @video.pro == true  %>
-                  <div class="verified">
-                    <i class="fas fa-key"><p>Vidéo professionnalisante</p></i>
-                  </div>
-                <% end %>
-              </div>
-              <hr>
-              <div id="accordion">
-                <div class="card-header" id="headingTwo">
-                  <h5 class="mb-0">
-                    <button class="btn btn-link collapsed" data-toggle="collapse" data-target="#collapseTwo" aria-expanded="false" aria-controls="collapseTwo">
-                      Description ...
-                    </button>
-                  </h5>
-                </div>
-                <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordion">
-                  <div class="card-body video-content">
-                    <%= @video.description %>
-                  </div>
-                </div>
-              </div>
-              <hr>
+          <div class="large-video-card-wrapper light-star-card-no-hover">
+            <div class="iframe-container">
+              <iframe src="https://www.youtube.com/embed/<%= @video.youtube_id %>?rel=0" allowfullscreen></iframe>
             </div>
+              <div class="bottom-video">
+                <div class="video-details-wrapper">
+
+                  <div><i class="far fa-comments"> 12</i></div>
+                  <div><i class="far fa-eye"> 2</i></div>
+                  <div class="icon-wrapper"><i class="far fa-clock"></i> <%= @video.duration_seconds / 60 %>:<%= "%.2i" %(@video.duration_seconds % 60) %>
+                  </div>
+                </div>
+                <div class="verified-wrapper">
+                  <% if @video.verified == true  %>
+                    <div class="verified">
+                      <i class="fas fa-check-circle"><p>Vidéo vérifiée</p></i>
+                    </div>
+                  <% end %>
+                  <% if @video.introduction == true  %>
+                    <div class="verified">
+                      <i class="fas fa-info-circle"><p>Vidéo d'introduction</p></i>
+
+                    </div>
+                  <% end %>
+                  <% if  @video.pro == true  %>
+                    <div class="verified">
+                      <i class="fas fa-key"><p>Vidéo professionnalisante</p></i>
+                    </div>
+                  <% end %>
+                </div>
+                <div id="accordion">
+                  <div class="card-header" id="headingTwo">
+                    <h5 class="mb-0">
+                      <button class="btn btn-link collapsed" data-toggle="collapse" data-target="#collapseTwo" aria-expanded="false" aria-controls="collapseTwo">
+                        Description ...
+                      </button>
+                    </h5>
+                  </div>
+                  <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordion">
+                    <div class="card-body video-content">
+                      <%= @video.description %>
+                    </div>
+                  </div>
+                </div>
+              </div>
+          </div>
           </div>
         </div>
       </div>
